+++ components/freshservice/src/inputs/agents/listAgents.ts
@@ -2,13 +2,11 @@ import {
   additionalQueryParams,
   connection,
   fetchAll,
-  page,
-  perPage,
+  pagination,
 } from "../common";
 export const listAgentsInputs = {
   connection,
   fetchAll,
-  perPage,
-  page,
+  pagination,
   additionalQueryParams,
 };