+++ components/freshservice/src/inputs/assets/updateAsset.ts
@@ -1,4 +1,4 @@
-import { input } from "@prismatic-io/spectral";
+import { input, structuredObjectInput } from "@prismatic-io/spectral";
 import { cleanNumberInput, cleanStringInput } from "../../util";
 import { connection } from "../common";
 import {
@@ -15,26 +15,35 @@ import {
   name,
   usageType,
 } from "./common";
+const additionalFields = structuredObjectInput({
+  label: "Additional Fields",
+  required: false,
+  comments:
+    "Additional optional fields: includes Name, Asset Tag, Impact, Usage Type, and Description.",
+  inputs: {
+    name: input({
+      ...name,
+      required: false,
+      clean: cleanStringInput,
+    }),
+    assetTag,
+    impact,
+    usageType,
+    description,
+  },
+});
 export const updateAssetInputs = {
   connection,
   assetDisplayId,
-  name: input({
-    ...name,
-    required: false,
-    clean: cleanStringInput,
-  }),
   assetTypeId: input({
     ...assetTypeId,
     required: false,
     clean: cleanNumberInput,
   }),
-  assetTag,
-  impact,
-  usageType,
-  description,
   locationId,
   agentId,
   departmentId,
   groupId,
+  additionalFields,
   assetsAdditionalFields,
 };