+++ components/freshservice/src/inputs/software/updateSoftware.ts
@@ -1,4 +1,4 @@
-import { input } from "@prismatic-io/spectral";
+import { input, structuredObjectInput } from "@prismatic-io/spectral";
 import { cleanStringInput } from "../../util";
 import { connection } from "../common";
 import {
@@ -13,28 +13,37 @@ import {
   source,
   status,
 } from "./common";
+const additionalFields = structuredObjectInput({
+  label: "Additional Fields",
+  required: false,
+  comments:
+    "Additional optional fields: includes Name, Description, Application Type, Status, Notes, Category, and Source.",
+  inputs: {
+    name: input({
+      ...name,
+      required: false,
+      clean: cleanStringInput,
+    }),
+    description: input({
+      ...description,
+      required: false,
+      clean: cleanStringInput,
+    }),
+    applicationType: input({
+      ...applicationType,
+      required: false,
+      clean: cleanStringInput,
+    }),
+    status,
+    notes,
+    category,
+    source,
+  },
+});
 export const updateSoftwareInputs = {
   connection,
   applicationId,
-  name: input({
-    ...name,
-    required: false,
-    clean: cleanStringInput,
-  }),
-  description: input({
-    ...description,
-    required: false,
-    clean: cleanStringInput,
-  }),
-  applicationType: input({
-    ...applicationType,
-    required: false,
-    clean: cleanStringInput,
-  }),
-  status,
   managedById,
-  notes,
-  category,
-  source,
+  additionalFields,
   softwareAdditionalFields,
 };