+++ components/freshservice/src/inputs/tickets/listTickets.ts
@@ -2,15 +2,13 @@ import {
   additionalQueryParams,
   connection,
   fetchAll,
-  page,
-  perPage,
+  pagination,
 } from "../common";
 import { filter } from "./common";
 export const listTicketsInputs = {
   connection,
   filter,
   fetchAll,
-  perPage,
-  page,
+  pagination,
   additionalQueryParams,
 };