+++ components/freshservice/src/inputs/tickets/updateTicket.ts
@@ -1,4 +1,4 @@
-import { input } from "@prismatic-io/spectral";
+import { input, structuredObjectInput } from "@prismatic-io/spectral";
 import { cleanNumberInput } from "../../util";
 import { connection } from "../common";
 import {
@@ -9,12 +9,21 @@ import {
   ticketId,
   ticketsAdditionalFields,
 } from "./common";
+const additionalFields = structuredObjectInput({
+  label: "Additional Fields",
+  required: false,
+  comments:
+    "Additional optional fields: includes Priority, Status, Source, and Bypass Mandatory.",
+  inputs: {
+    priority: input({ ...priority, required: false, clean: cleanNumberInput }),
+    status: input({ ...status, required: false, clean: cleanNumberInput }),
+    source: input({ ...source, required: false, clean: cleanNumberInput }),
+    bypassMandatory,
+  },
+});
 export const updateTicketInputs = {
   connection,
   ticketId: input({ ...ticketId, comments: "ID of the ticket to update." }),
-  priority: input({ ...priority, required: false, clean: cleanNumberInput }),
-  status: input({ ...status, required: false, clean: cleanNumberInput }),
-  source: input({ ...source, required: false, clean: cleanNumberInput }),
-  bypassMandatory,
+  additionalFields,
   ticketsAdditionalFields,
 };